% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_summary_opensafely.R
\name{advanced_summary_opensafely}
\alias{advanced_summary_opensafely}
\title{Advanced Statistical Summary for OpenSAFELY-like Datasets}
\usage{
advanced_summary_opensafely(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A list with detailed summaries
}
\description{
Advanced Statistical Summary for OpenSAFELY-like Datasets
}
