% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_opensafely.R
\name{summary_opensafely}
\alias{summary_opensafely}
\title{Print full summary of health dataset in console}
\usage{
summary_opensafely(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Print full summary of health dataset in console
}
